Vika

+++ src/routes/comments.py
@@ -1,8 +1,9 @@
-from fastapi import APIRouter, Request, Depends
+from fastapi import APIRouter, Request, Depends, HTTPException, status
 
 from src.database.db import get_db
 from src.limiter import limiter
 from src.repository import comments as comments_repository
+from src.database.models import UserRole
 from src.schemas import CommentResponse
 from src.services.auth import auth_service
 
@@ -21,13 +22,11 @@ async def add_comment(
 ):
     """
     Adding comment to db.
-
     :param request: FastAPI Request instance.
     :param text: Comment text.
     :param image_id: Image id number.
     :param db: The SQLAlchemy Session instance.
     :param user: Get the current authenticated user.
-
     :return:
     """
     return await comments_repository.add_comment(
@@ -46,15 +45,19 @@ async def edit_comment(
 ):
     """
     Editing comment in db.
-
     :param request: FastAPI Request instance.
     :param text: Comment text.
     :param comment_id: Comment id number.
     :param db: The SQLAlchemy Session instance.
     :param user: Get the current authenticated user.
-
     :return:
     """
+    if not text.strip():
+        raise HTTPException(
+        status_code=status.HTTP_400_BAD_REQUEST,
+        detail="Comment text cannot be empty or contain only whitespace",
+        )
+    
     return await comments_repository.edit_comment(
         text=text, comment_id=comment_id, user=user, db=db
     )
@@ -70,14 +73,16 @@ async def delete_comment(
 ):
     """
     Deleting comment in db.
-
     :param request: FastAPI Request instance.
     :param comment_id: Comment id number.
     :param db: The SQLAlchemy Session instance.
     :param user: Get the current authenticated user.
-
     :return:
     """
+    if user.role not in (UserRole.admin.value, UserRole.moderator.value):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+        detail="Only administrators and moderators can delete comments")
+    
     return await comments_repository.delete_comment(
         comment_id=comment_id, user=user, db=db
     )
@@ -93,12 +98,10 @@ async def get_comments_by_image_id(
 ):
     """
     Getting comments by image id.
-
     :param request: FastAPI Request instance.
     :param image_id: Image id number.
     :param db: The SQLAlchemy Session instance.
     :param user: Get the current authenticated user.
-
     :return:
     """
     return await comments_repository.get_comments_by_image_id(image_id=image_id, db=db)
